requirements.txt
mnemonic==0.20
bip-utils==2.9.0
requests==2.31.0
base58==2.1.1
PyNaCl==1.5.0 